added class = "gif"
and made that CSS display:none so it doesn't show and only the
question's GIF shows.

This is not the best way to write the code for this HW because our goal
is Object Oriented Programming and Pavan suggests I write it in OOP
way. So that's it for this one, I'm going to write a new one.